Added TileMap Editor for level design
To access the Tile map - Click GameObject and at the bottom should be
"Tile Map". A tile map is created with multiple inspector options. An
image or texture can be selected. Click the "Window" tab and select
"Tile Picker" to view the different tiles to use.

When editing tiles/textures Holding shift keyon a tile will paint the
selected file. Holding Alt key will erase tiles. There is a clear all
tiles option in the inspector.

In Addition to this I have moved some files around to sort scripts and
our scenes better. Created a test scene to work with and test out
different tiles.